+++ srcs/physician/pressurer.hpp
@@ -18,37 +18,37 @@ public:
     void solvePressure(void);
 private:
     struct Acell {
-        double diag;
-        double plusi;
-        double plusj;
+        float diag;
+        float plusi;
+        float plusj;
     };
 
     void updateVelocity(void);
     void calcNegativeDivergence(void);
     void calcA();
-    void applyPrecon(double (&r)[GRID_WIDTH][GRID_HEIGHT], double (&res)[GRID_WIDTH][GRID_HEIGHT]);
+    void applyPrecon(float (&r)[GRID_WIDTH][GRID_HEIGHT], float (&res)[GRID_WIDTH][GRID_HEIGHT]);
     void calcPrecon();
     void PCG(void);
-    void cpyVect(double (&original)[GRID_WIDTH][GRID_HEIGHT], double (&res)[GRID_WIDTH][GRID_HEIGHT]);
-    void applyA(double (&s)[GRID_WIDTH][GRID_HEIGHT], double (&res)[GRID_WIDTH][GRID_HEIGHT]);
+    void cpyVect(float (&original)[GRID_WIDTH][GRID_HEIGHT], float (&res)[GRID_WIDTH][GRID_HEIGHT]);
+    void applyA(float (&s)[GRID_WIDTH][GRID_HEIGHT], float (&res)[GRID_WIDTH][GRID_HEIGHT]);
 
     /* utilitaire de vect */
-    void addVect(double (&a)[GRID_WIDTH][GRID_HEIGHT], double (&b)[GRID_WIDTH][GRID_HEIGHT], double (&res)[GRID_WIDTH][GRID_HEIGHT]);
-    void multScalVect(double coef, double (&vect)[GRID_WIDTH][GRID_HEIGHT], double (&res)[GRID_WIDTH][GRID_HEIGHT]);
-    double dotProduct(double (&a)[GRID_WIDTH][GRID_HEIGHT], double (&b)[GRID_WIDTH][GRID_HEIGHT]);
-    double normeVect(double (&vect)[GRID_WIDTH][GRID_HEIGHT]);
-    void bzeroVect(double (&vect)[GRID_WIDTH][GRID_HEIGHT]);
+    void addVect(float (&a)[GRID_WIDTH][GRID_HEIGHT], float (&b)[GRID_WIDTH][GRID_HEIGHT], float (&res)[GRID_WIDTH][GRID_HEIGHT]);
+    void multScalVect(float coef, float (&vect)[GRID_WIDTH][GRID_HEIGHT], float (&res)[GRID_WIDTH][GRID_HEIGHT]);
+    float dotProduct(float (&a)[GRID_WIDTH][GRID_HEIGHT], float (&b)[GRID_WIDTH][GRID_HEIGHT]);
+    float normeVect(float (&vect)[GRID_WIDTH][GRID_HEIGHT]);
+    void bzeroVect(float (&vect)[GRID_WIDTH][GRID_HEIGHT]);
 
     struct Acell   A[GRID_WIDTH][GRID_HEIGHT];
-    double         precon[GRID_WIDTH][GRID_HEIGHT];
-    double         z[GRID_WIDTH][GRID_HEIGHT];
-    double         s[GRID_WIDTH][GRID_HEIGHT];
-    double         b[GRID_WIDTH][GRID_HEIGHT];
-    double         p[GRID_WIDTH][GRID_HEIGHT];
-    double         r[GRID_WIDTH][GRID_HEIGHT];
+    float         precon[GRID_WIDTH][GRID_HEIGHT];
+    float         z[GRID_WIDTH][GRID_HEIGHT];
+    float         s[GRID_WIDTH][GRID_HEIGHT];
+    float         b[GRID_WIDTH][GRID_HEIGHT];
+    float         p[GRID_WIDTH][GRID_HEIGHT];
+    float         r[GRID_WIDTH][GRID_HEIGHT];
 
     /* for the applyPrecon */
-    double         q[GRID_WIDTH][GRID_HEIGHT];
+    float         q[GRID_WIDTH][GRID_HEIGHT];
 };
 }
